requirements.txt
streamlit
requests
openai
python-dotenv
fastapi
uvicorn
httpx
plotly